Move uninstall file selection logic to req_uninstall.py
